[F] Add resuable Debugger component
