+++ client/src/components/global/Developer/index.js
@@ -1,0 +1,5 @@
+import Debugger from './Debugger';
+
+export default {
+  Debugger
+};
