+++ client/src/containers/backend/Form/Form.js
@@ -2,10 +2,10 @@ import React, { PureComponent, PropTypes } from 'react';
 import { withRouter } from 'react-router';
 import { connect } from 'react-redux';
 import { entityEditorActions, entityStoreActions } from 'actions';
+import { Developer } from 'components/global';
 import { bindActionCreators } from 'redux';
 import { entityUtils } from 'utils';
 import get from 'lodash/get';
-import JSONTree from 'react-json-tree'
 
 const { select } = entityUtils;
 const { request, flush } = entityStoreActions;
@@ -160,51 +160,9 @@ class FormContainer extends PureComponent {
   }
 
   renderDebugger() {
-    if (!__DEVELOPMENT__) return null;
     if (!this.props.debug) return null;
-    const theme = {
-      scheme: 'green screen',
-      author: 'chris kempson (http://chriskempson.com)',
-      base00: '#001100',
-      base01: '#003300',
-      base02: '#005500',
-      base03: '#007700',
-      base04: '#009900',
-      base05: '#00bb00',
-      base06: '#00dd00',
-      base07: '#00ff00',
-      base08: '#007700',
-      base09: '#009900',
-      base0A: '#007700',
-      base0B: '#00bb00',
-      base0C: '#005500',
-      base0D: '#009900',
-      base0E: '#00bb00',
-      base0F: '#005500'
-    }
-    const advanced = {
-      extend: theme,
-      tree: {
-        padding: 20,
-        margin: 0
-      }
-    }
-    return (
-      <div style={{
-        marginBottom: 15
-      }}>
-        <div style={{
-          backgroundColor: "rgb(0, 17, 0)",
-          color: "rgb(0, 153, 0)",
-          margin: 0,
-          padding: "10px 20px",
-          display: "inline-block"
-        }}>Form Debugger</div>
-        <JSONTree hideRoot={true} theme={advanced} invertTheme={false} data={this.props.session} />
-      </div>
-    )
+    return <Developer.Debugger object={this.props.session} />;
   }
-
 
   render() {
     if (!this.props.session) return null;
